refactor(gui): split GUI into focused sub-packages
Decompose the monolithic GUI module into smaller, single-responsibility
sub-packages and modules:

- controllers/: extract ThemeManager, NavigationManager, and
  UpdateManager from GUIController; controller.py now delegates via
  thin stubs, preserving the public API
- utils/: consolidate git, io, sequence, ui, and version_check modules
  into data_helpers, gui_helpers, and system; remove the old files
- views/input/primer/: reorganise primer input into a sub-package with
  clipboard, coordinator, action_controller, header, info_panel,
  layout_manager, toolbar, and validation modules
- views/input/template/: reorganise template input into a sub-package
  with casing, file_manager, and formatter modules
- views/input/: extract shared event and layout logic into events.py
  and layout.py; slim down input_view.py
- views/settings/: rename base_score_tile.py to score_table.py

Update all imports across the codebase and tests accordingly.
