+++ src/amplifyp/gui/views/input/primer/input.py
@@ -19,29 +19,25 @@ from __future__ import annotations
 
 import logging
 from collections.abc import Callable
-from typing import Any, cast
+from typing import cast
 
 import flet as ft
 
 from amplifyp.gui.colours import GUIColours
 from amplifyp.gui.settings import GUISettings
 from amplifyp.gui.user_data import GUIInput
-from amplifyp.gui.utils.sequence import clean_sequence
-from amplifyp.gui.utils.ui import NotificationHelper
+from amplifyp.gui.utils.data_helpers import clean_sequence
+from amplifyp.gui.utils.gui_helpers import NotificationHelper
 
-from .primer_action_controller import PrimerActionController
-from .primer_file_manager import PrimerFileManager
-from .primer_header import PrimerHeader
-from .primer_info_panel import PrimerInfoPanel
-from .primer_layout_manager import PrimerLayoutManager
-from .primer_list import PrimerList
-from .primer_row import PrimerRow
-from .primer_toolbar import PrimerToolbar
-from .primer_validation import (
-    get_duplicate_primer_indices,
-    reconcile_primer_states,
-    validate_primers,
-)
+from .action_controller import PrimerActionController
+from .coordinator import PrimerCoordinator
+from .file_manager import PrimerFileManager
+from .header import PrimerHeader
+from .info_panel import PrimerInfoPanel
+from .layout_manager import PrimerLayoutManager
+from .list import PrimerList
+from .row import PrimerRow
+from .toolbar import PrimerToolbar
 
 logger = logging.getLogger(__name__)
 
@@ -61,19 +57,7 @@ class PrimerInput(ft.Container):  # type: ignore[misc]
         clear_primers_callback: Callable[[ft.Event | None], None],
         delete_selected_callback: Callable[[ft.Event | None], None],
     ) -> None:
-        """Initialise the PrimerInput component.
-
-        Args:
-            page: The Flet page instance for file picker and notifications.
-            settings: Application GUI settings instance.
-            input_data: Central GUI input state object.
-            on_change_handler: Callback invoked when primer content changes.
-            handle_field_focus: Callback for field focus events.
-            handle_field_blur: Callback for field blur events.
-            handle_field_submit: Callback for field submit events.
-            clear_primers_callback: Callback to clear all primers.
-            delete_selected_callback: Callback to delete selected primers.
-        """
+        """Initialise the PrimerInput component."""
         super().__init__(expand=5)
         self.app_page = page
         self.settings = settings
@@ -87,6 +71,7 @@ class PrimerInput(ft.Container):  # type: ignore[misc]
 
         self.layout_manager = PrimerLayoutManager(self)
         self.action_controller = PrimerActionController(self)
+        self.coordinator = PrimerCoordinator(self)
 
         self.focused_primer_index: int | None = None
         self.selected_indices: set[int] = set()
@@ -112,12 +97,11 @@ class PrimerInput(ft.Container):  # type: ignore[misc]
             on_divider_pan=self.layout_manager.on_primer_divider_pan,
             on_divider_pan_end=self.layout_manager.on_primer_divider_pan_end,
             name_column_width=self.name_column_width,
-            on_add_primer=self._header_add_click,
-            on_delete_primer=self._header_delete_click,
-            on_move_primer_up=self._header_up_click,
-            on_move_primer_down=self._header_down_click,
+            on_add_primer=self.action_controller.header_add_click,
+            on_delete_primer=self.action_controller.header_delete_click,
+            on_move_primer_up=self.action_controller.header_up_click,
+            on_move_primer_down=self.action_controller.header_down_click,
         )
-        # Compatibility links
         self.all_primers_checkbox = self.primer_header.all_primers_checkbox
         self.primers_header = self.primer_header.header_row
         self.primers_header_container = self.primer_header
@@ -140,7 +124,6 @@ class PrimerInput(ft.Container):  # type: ignore[misc]
             on_copy=self._copy_primers_click,
             on_paste=self._paste_primers_click,
         )
-        # Compatibility links
         self.clear_primers_button = self.primer_toolbar.clear_button
         self.delete_selected_button = self.primer_toolbar.delete_selected_button
 
@@ -167,7 +150,7 @@ class PrimerInput(ft.Container):  # type: ignore[misc]
             settings=self.settings, font_family=font_family
         )
 
-        # Primer List Container (stored for dynamic repositioning)
+        # Primer List Container
         self.primer_list_container = ft.Container(
             content=ft.Column(
                 [
@@ -183,7 +166,7 @@ class PrimerInput(ft.Container):  # type: ignore[misc]
             padding=0,
         )
 
-        # Title row (stored for dynamic repositioning)
+        # Title row
         self.primer_title_row = ft.ResponsiveRow(
             [
                 ft.Row(
@@ -214,18 +197,10 @@ class PrimerInput(ft.Container):  # type: ignore[misc]
             spacing=5,
         )
 
-        # Position info panel based on setting
         self._reposition_info_panel()
 
     def _reposition_info_panel(self) -> None:
-        """Reposition the primer info panel based on the current setting.
-
-        Moves the info panel to either the top (after the toolbar row)
-        or the bottom (before the error banner) of the primer list area.
-
-        Rebuilds the entire Column so Flet detects the structural change
-        and re-renders the new order.
-        """
+        """Reposition the primer info panel based on the current setting."""
         position = str(
             self.settings.get("primer_info_panel_position", "bottom")
         ).lower()
@@ -253,22 +228,15 @@ class PrimerInput(ft.Container):  # type: ignore[misc]
                 expand=True,
                 spacing=5,
             )
-        # Trigger Flet to re-render the changed content
         try:
             self.update()
         except RuntimeError:
             pass
 
     def reposition_info_panel(self) -> None:
-        """Public method to reposition the info panel.
-
-        Call this when the setting changes to ensure the panel is
-        immediately repositioned. The internal update is handled by
-        _reposition_info_panel.
-        """
+        """Public method to reposition the info panel."""
         self._reposition_info_panel()
 
-    # Delegate properties of the info panel for test and backward compatibility
     @property
     def info_header(self) -> ft.Container:
         """Get the primer info header container."""
@@ -299,142 +267,32 @@ class PrimerInput(ft.Container):  # type: ignore[misc]
         """Get the primer info dimer text control."""
         return self.primer_info_panel.info_dimer_text
 
-    def _extract_primer_data_from_ui(self) -> list[dict[str, Any]]:
-        """Extract primer data from UI controls.
-
-        Returns:
-            A list of dicts with keys 'name', 'seq', 'active', 'container',
-            and 'checkbox' for each primer row in the UI.
-        """
-        ui_primers = []
-        for row in self.primers_list.controls:
-            if not isinstance(row, PrimerRow):
-                continue
-
-            name_val = str(row.name_field.value or "").strip()
-            seq_val = clean_sequence(str(row.seq_field.value or ""))
-            is_active = bool(row.checkbox.value)
-
-            ui_primers.append(
-                {
-                    "name": name_val,
-                    "seq": seq_val,
-                    "active": is_active,
-                    "container": row,
-                    "checkbox": row.checkbox,
-                }
-            )
-        return ui_primers
-
     def sync_to_state(
         self, rebuild_if_needed: bool = False, skip_extract: bool = False
     ) -> bool:
-        """Sync current UI controls back to the central state.
-
-        Args:
-            rebuild_if_needed: If True, triggers a reconciliation / UI rebuild.
-            skip_extract: If True, skips UI extraction and uses existing
-                state directly.
-        """
-        if skip_extract:
-            ui_primers = self.input_data.primers
-        else:
-            ui_primers = self._extract_primer_data_from_ui()
-
-        auto_activate_new = self.settings.get(
-            "auto_activate_new_valid_primer", False
-        )
-        primers = reconcile_primer_states(
-            ui_primers,
-            self.input_data.primers,
-            auto_activate_new=auto_activate_new,
-        )
-
-        # Update checkbox values in-place on UI controls if they
-        # were updated during reconciliation.  The subsequent
-        # _update_header_checkbox_state() → page.update() call flushes
-        # all dirty controls in a single batch, which is the reliable way
-        # to propagate aria-checked changes to the Flutter semantics layer
-        # in web mode.  An earlier individual checkbox.update() would race
-        # with that full-page flush and leave the semantics stale.
-        for reconciled_p, ui_p in zip(primers, ui_primers, strict=True):
-            checkbox = ui_p.get("checkbox")
-            if checkbox:
-                checkbox.value = reconciled_p["active"]
-
-        ignore_inactive_name_dup = self.settings.get(
-            "ignore_inactive_name_dup_warn", True
-        )
-        ignore_inactive_seq_dup = self.settings.get(
-            "ignore_inactive_seq_dup_warn", True
-        )
-
-        dup_indices = get_duplicate_primer_indices(
-            ui_primers, ignore_inactive_name_dup, ignore_inactive_seq_dup
-        )
-        for p in ui_primers:
-            container = p.get("container")
-            if container is None:
-                continue
-            c_idx = container.data
-            is_dup = c_idx in dup_indices
-            new_color = GUIColours.DUPLICATE_BG if is_dup else None
-            if container.bgcolor != new_color:
-                container.bgcolor = new_color
-
-        # Run background primer construction/validation
-        new_validation_errors = validate_primers(
-            primers, ignore_inactive_name_dup, ignore_inactive_seq_dup
-        )
-
-        self.input_data.primers = primers
-        self.validation_errors = new_validation_errors
-        if rebuild_if_needed:
-            self.update_ui()
-        else:
-            # Update error status and duplicate highlights in-place on
-            # existing controls
-            for idx, row in enumerate(self.primers_list.controls):
-                if isinstance(row, PrimerRow) and idx < len(
-                    new_validation_errors
-                ):
-                    row.set_error(new_validation_errors[idx])
-                    row.update_tm(self.settings)
-            self._update_row_highlights()
-            self._update_header_checkbox_state()
-            self._update_delete_button_disabled_state()
-
-        return rebuild_if_needed
+        """Sync current UI controls back to the central state."""
+        return self.coordinator.sync_to_state(rebuild_if_needed, skip_extract)
 
     def update_ui(self) -> None:
-        """Update Flet UI controls to match the central state.
-
-        Rebuilds the primer list and updates the delete button disabled
-        state based on current selections.
-        """
-        # Recreate the header to make sure controls match settings and indices
+        """Update Flet UI controls to match the central state."""
         self.primer_header = PrimerHeader(
             settings=self.settings,
             on_toggle_all=self._on_toggle_all_primers,
             on_divider_pan=self.layout_manager.on_primer_divider_pan,
             on_divider_pan_end=self.layout_manager.on_primer_divider_pan_end,
             name_column_width=self.name_column_width,
-            on_add_primer=self._header_add_click,
-            on_delete_primer=self._header_delete_click,
-            on_move_primer_up=self._header_up_click,
-            on_move_primer_down=self._header_down_click,
+            on_add_primer=self.action_controller.header_add_click,
+            on_delete_primer=self.action_controller.header_delete_click,
+            on_move_primer_up=self.action_controller.header_up_click,
+            on_move_primer_down=self.action_controller.header_down_click,
         )
         self.all_primers_checkbox = self.primer_header.all_primers_checkbox
         self.primers_header = self.primer_header.header_row
         self.primers_header_container = self.primer_header
 
-        # Reposition info panel based on current setting
         self._reposition_info_panel()
-
-        # Update the info panel to reflect any new settings or focus state
         self._update_primer_info_panel()
 
-        # Replace the header in the UI container controls
         inner_column = cast(ft.Column, self.primer_list_container.content)
         inner_column.controls[0] = self.primer_header
         try:
@@ -453,22 +311,11 @@ class PrimerInput(ft.Container):  # type: ignore[misc]
             pass
 
     def _update_delete_button_disabled_state(self) -> None:
-        """Update disabled state of the delete button based on selection.
-
-        The delete button is enabled only when at least one primer is
-        active (selected).
-        """
+        """Update disabled state of the delete button based on selection."""
         self.delete_selected_button.disabled = not self.selected_indices
 
     def _on_toggle_all_primers(self, e: ft.Event[ft.Checkbox]) -> None:
-        """Toggle all primers active/inactive based on tri-state checkbox.
-
-        Handles the three checkbox states (True, False, None/tristate)
-        to determine the target active state for all primers.
-
-        Args:
-            e: The Flet control event triggered by the checkbox change.
-        """
+        """Toggle all primers active/inactive based on tri-state checkbox."""
         primers = self.input_data.primers
         if not primers:
             return
@@ -488,7 +335,6 @@ class PrimerInput(ft.Container):  # type: ignore[misc]
             else:
                 target_active = True
 
-        # Update checkbox values in-place on existing UI controls
         for row in self.primers_list.controls:
             if isinstance(row, PrimerRow) and row.data is not None:
                 idx = row.data
@@ -503,11 +349,7 @@ class PrimerInput(ft.Container):  # type: ignore[misc]
         self.on_change_handler(e)
 
     def _update_header_checkbox_state(self) -> None:
-        """Update the header checkbox to reflect the current primer states.
-
-        Sets the tri-state checkbox to True (all active), False (none
-        active), or None (mixed state) based on the current primer list.
-        """
+        """Update the header checkbox to reflect the current primer states."""
         primers = self.input_data.primers
         non_empty = [
             p
@@ -528,39 +370,16 @@ class PrimerInput(ft.Container):  # type: ignore[misc]
             self.app_page.update()
 
     def _get_duplicate_indices(self) -> set[int]:
-        """Find indices of primers with duplicate names or sequences.
-
-        Returns:
-            Set of indices corresponding to primers with duplicate names
-            or sequences in the central state.
-        """
-        ignore_inactive_name_dup = self.settings.get(
-            "ignore_inactive_name_dup_warn", True
-        )
-        ignore_inactive_seq_dup = self.settings.get(
-            "ignore_inactive_seq_dup_warn", True
-        )
-        return get_duplicate_primer_indices(
-            self.input_data.primers,
-            ignore_inactive_name_dup,
-            ignore_inactive_seq_dup,
-        )
+        """Find indices of primers with duplicate names or sequences."""
+        return self.coordinator.get_duplicate_indices()
 
     def _update_row_highlights(self) -> None:
-        """Update background colours of all row containers.
-
-        Highlights rows based on selection (focused primer) and
-        duplicates (by name or sequence).
-        """
+        """Update background colours of all row containers."""
         self.primers_list.update_row_highlights()
         self._update_header_buttons_state()
 
     def _update_primer_info_panel(self) -> None:
-        """Update the primer information panel based on the focused primer.
-
-        Displays Tm, sequence details, base counts, redundancy, and
-        dimer potential for the currently focused primer row.
-        """
+        """Update the primer information panel based on the focused primer."""
 
         def on_dismiss() -> None:
             """Clear focus when the primer info panel is dismissed."""
@@ -578,123 +397,30 @@ class PrimerInput(ft.Container):  # type: ignore[misc]
         )
 
     async def _load_primers_click(self, e: ft.Event | None) -> None:
-        """Open file picker to load primers from CSV/TSV file.
-
-        Delegates to the PrimerFileManager component.
-
-        Args:
-            e: The Flet control event triggered by the load button click.
-        """
+        """Open file picker to load primers from CSV/TSV file."""
         await self.file_manager.load_primers_click(e)
 
     async def _save_primers_click(self, e: ft.Event | None) -> None:
-        """Save active primers to a CSV file.
-
-        Delegates to the PrimerFileManager component.
-
-        Args:
-            e: The Flet control event triggered by the save button click.
-        """
+        """Save active primers to a CSV file."""
         await self.file_manager.save_primers_click(e)
 
     def _show_notification(self, message: str) -> None:
-        """Show a notification message.
-
-        Args:
-            message: The notification message text to display.
-        """
+        """Show a notification message."""
         if not hasattr(self, "_notification_helper"):
             self._notification_helper = NotificationHelper(self.app_page)
         self._notification_helper.show_message(message)
 
     async def _copy_primers_click(self, e: ft.Event | None) -> None:
         """Copy selected or focused primers to clipboard in TSV format."""
-        primers = self.input_data.primers
-        selected_primers = (
-            [
-                primers[i]
-                for i in sorted(self.selected_indices)
-                if 0 <= i < len(primers)
-            ]
-            if self.selected_indices
-            else []
-        )
+        from .clipboard import copy_primers_click
 
-        # Fallback to focused primer if no selected rows
-        if not selected_primers and self.focused_primer_index is not None:
-            if 0 <= self.focused_primer_index < len(primers):
-                selected_primers = [primers[self.focused_primer_index]]
-
-        if not selected_primers:
-            self._show_notification(
-                "No primers highlighted or focused to copy."
-            )
-            return
-
-        lines = []
-        for p in selected_primers:
-            name = str(p.get("name", "")).strip()
-            seq = str(p.get("seq", "")).strip()
-            lines.append(f"{name}\t{seq}")
-
-        tsv_text = "\n".join(lines)
-        await ft.Clipboard().set(tsv_text)
-        self._show_notification(
-            f"Copied {len(selected_primers)} primer(s) to clipboard."
-        )
+        await copy_primers_click(self, e)
 
     async def _paste_primers_click(self, e: ft.Event | None) -> None:
         """Paste primers from clipboard starting at focused index or end."""
-        try:
-            clipboard_text = await ft.Clipboard().get()
-        except Exception as ex:
-            logger.warning("Failed to access clipboard: %s", ex)
-            self._show_notification(
-                "Unable to read clipboard. Try pasting directly into a "
-                "primer field using Ctrl+V."
-            )
-            return
-        if not clipboard_text:
-            self._show_notification("Clipboard is empty.")
-            return
+        from .clipboard import paste_primers_click
 
-        from .primer_clipboard import parse_primer_clipboard_text
-
-        parsed = parse_primer_clipboard_text(clipboard_text)
-        if not parsed:
-            self._show_notification("No valid primers found in clipboard.")
-            return
-
-        primers = self.input_data.primers
-        valid_selected = {
-            i for i in self.selected_indices if 0 <= i < len(primers)
-        }
-        if valid_selected:
-            insert_idx = max(valid_selected) + 1
-        else:
-            insert_idx = len(primers)
-
-        # Replace single empty row
-        if (
-            len(primers) == 1
-            and not primers[0].get("name")
-            and not primers[0].get("seq")
-        ):
-            primers.clear()
-            insert_idx = 0
-
-        for i, new_p in enumerate(parsed):
-            new_p["active"] = False
-            primers.insert(insert_idx + i, new_p)
-
-        self.selected_indices = set(range(insert_idx, insert_idx + len(parsed)))
-        self.focused_primer_index = insert_idx
-
-        self.sync_to_state(rebuild_if_needed=True, skip_extract=True)
-        if self.on_change_handler is not None:
-            self.on_change_handler(None)
-
-        self._show_notification(f"Pasted {len(parsed)} primer(s).")
+        await paste_primers_click(self, e)
 
     def _update_header_buttons_state(self) -> None:
         """Update enabled/disabled state of header action buttons."""
@@ -721,38 +447,3 @@ class PrimerInput(ft.Container):  # type: ignore[misc]
                 self.primer_header.update()
         except RuntimeError:
             logger.debug("Header page detached, skipping update")
-
-    def _header_add_click(self, e: ft.Event | None) -> None:
-        """Handle header Add button click."""
-        num_primers = len(self.input_data.primers)
-        if self.selected_indices:
-            idx = min(max(self.selected_indices), num_primers - 1)
-        elif self.focused_primer_index is not None:
-            idx = min(self.focused_primer_index, num_primers - 1)
-        else:
-            idx = num_primers - 1
-
-        self.selected_indices = {idx + 1}
-        self.focused_primer_index = idx + 1
-        self.action_controller.on_add_primer_row(idx)
-
-    def _header_delete_click(self, e: ft.Event | None) -> None:
-        """Handle header Delete button click."""
-        if self.selected_indices:
-            self.action_controller.delete_primers(self.selected_indices.copy())
-            self._update_header_buttons_state()
-
-    def _header_up_click(self, e: ft.Event | None) -> None:
-        """Handle header Move Up button click."""
-        if self.selected_indices and min(self.selected_indices) > 0:
-            self.action_controller.move_primers(self.selected_indices, -1)
-            self._update_header_buttons_state()
-
-    def _header_down_click(self, e: ft.Event | None) -> None:
-        """Handle header Move Down button click."""
-        if (
-            self.selected_indices
-            and max(self.selected_indices) < len(self.input_data.primers) - 1
-        ):
-            self.action_controller.move_primers(self.selected_indices, 1)
-            self._update_header_buttons_state()
