+++ src/amplifyp/gui/views/pcr/primer_drawing.py
@@ -449,7 +449,9 @@ class ReplicationContextCard(DismissibleDetailCard):
         L = len(conf.primer)
         N = len(conf.template)
 
-        from amplifyp.gui.utils.sequence import create_overlapped_sequence_view
+        from amplifyp.gui.utils.data_helpers import (
+            create_overlapped_sequence_view,
+        )
 
         top_line, primer_line, bottom_line = format_context_lines(
             primer_name=primer_name,
