+++ src/amplifyp/gui/views/settings/primer_list_tile.py
@@ -13,7 +13,9 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-"""PrimerListTile component for Flet settings view."""
+"""PrimerListTile expansion tile component for settings view."""
+
+from __future__ import annotations
 
 from collections.abc import Callable
 from typing import Any
@@ -22,7 +24,7 @@ import flet as ft
 
 from amplifyp.gui.colours import GUIColours
 from amplifyp.gui.settings import GUISettings
-from amplifyp.gui.utils.ui import BorderedCheckbox
+from amplifyp.gui.utils.gui_helpers import BorderedCheckbox
 
 
 class PrimerListTile(ft.ExpansionTile):  # type: ignore[misc]
@@ -35,15 +37,7 @@ class PrimerListTile(ft.ExpansionTile):  # type: ignore[misc]
         on_change_handler: Callable[[ft.Event | None], None],
         header_size: int,
     ) -> None:
-        """Initialise the PrimerListTile.
-
-        Args:
-            settings: The settings object.
-            settings_map: A dictionary mapping setting keys to UI
-                components for population and retrieval.
-            on_change_handler: The handler to call when a setting changes.
-            header_size: The size of the expansion tile header text.
-        """
+        """Initialise the PrimerListTile."""
         self.settings = settings
         self.settings_map = settings_map
         self.on_change_handler = on_change_handler
