+++ tests/gui_state_test.py
@@ -513,7 +513,7 @@ def test_controller_load_save_only_affects_input() -> None:
         serialised_yaml = content
 
     with patch(
-        "amplifyp.gui.utils.io.save_and_write_file",
+        "amplifyp.gui.utils.data_helpers.save_and_write_file",
         new=AsyncMock(side_effect=mock_save_and_write_file),
     ):
         asyncio.run(controller.save_state(MagicMock()))
@@ -544,7 +544,7 @@ settings:
   dark_mode: false
 """
     with patch(
-        "amplifyp.gui.utils.io.pick_and_read_file",
+        "amplifyp.gui.utils.data_helpers.pick_and_read_file",
         new=AsyncMock(return_value=test_yaml),
     ):
         asyncio.run(controller.load_state(MagicMock()))
